requirements.txt
apache-airflow==1.10.1 # pyup: ignore
autoflake==1.3
black==19.3b0
flake8==3.7.7
google-api-python-client==1.7.8
j2cli==0.3.8
Jinja2==2.10.1
mypy==0.701
orderedset==2.0.1
parameterized==0.7.0
paramiko==2.4.2
pre-commit==1.16.1
pydevd==1.6.1
pylint==2.3.1
pytest-cov==2.7.1
safety==1.8.5
sshtunnel==0.1.4
